Added Pause script
I also added the inputManager script that store constants of all the possible inputs. (in future we can change this inputs so we can implement a binding system)
I also added the SceneSettings that store all the possible resolutions for the game. (in future we can implement a resolution system)